[Kernels][GPU] Do not fully inline the _check_error for GPU functions (#43055)
We were aggressively inlining the GPU check_error function which was
causing the compiler to stringify the error and location information at
each call site. This changes the code so that the error check is inlined
(to get location information) but the actual implementation is not.

GPU_ORIG_REV_ID: 25afa65a281e775a7091e153e1746732899e73c4