[Kernels] fix `apple_gemv` for unaligned K (#43122)
This fixes `apple_gemv` (really gevm) for the case where K is not a
multiple of the SIMD width. The scalar logic was trying to select
between two sources for B, but the order of operations was wrong and the
if/else ternary op was selecting between (a * b0) or (b1). The selection
logic for B is complicated, so the loads for a/b have been hoisted out
to simplify that and also fix the precedence bug.

Test cases were added to test_matmul.mojo for the failing test in
KERN-655 and to cover a larger GEVM too.

Resolves KERN-655.

LINALG_ORIG_REV_ID: 68cde67072be2b921fd8e3d882a85592c4ac259f