[MO][Support][MHA] Fix Permutation When Extracting Grouped Query Attention (#43117)
Stacked PRs:
 * __->__#43117

--- --- ---

### [MO][Support][MHA] Fix Permutation When Extracting Grouped Query
Attention

Fixes GRA-743.

The algorithm for extracting grouped key / values assumed the input was
layout BSHD and walked up a pattern.

When it encountered a transpose, it recorded the layout from before we
applied the transpose.

At the end, we have the layout of the initial grouped attention weights.
Unfortunately this may not be BSHD.

Previously, we take this layout and re-apply it as a permutation. This
sometimes works because a lot of permutations of size 4 when applied to
themselves return the identity permutation. However it is not true in
general. Instead, we must figure out how to translate the given
permutation back to the identity permutation representing BSHD.

Add a helper method for doing this and update some tests.

MODULAR_ORIG_COMMIT_REV_ID: 999652ae877b7fd6bcd309b87fe71cc222ff0c4d