[KGEN] Introduce LLVMModuleAndContext and pipe through ObjectCompiler (#43150)
One of the performance problems with the ObjectCompiler is there is a
lot of unnecesary writing and reading of IR from bitcode/bytecode, both
MLIR and LLVMIR. Sometimes we write a file to bytecode, read it back,
only to write it to bytecode again for caching.

Another problem is we unecessary roundtrip an LLVM module through
bitcode. We do this so that we can thread LLVM safely, but this is only
necessary if any splitting was involved. This PR introduces
LLVMModuleAndContext, a helper type that contains an LLVMContext and a
single LLVM module paired together. This also helps avoid lifetime
issues where the context goes out of scope before the module destructs.

Push this through the ObjectCompiler APIs, which cleans up a lot of
code, and also move the bitcode roundtripping into the module splitters,
which is where it actually matters. This allows us to abstract away when
it is safe to parallelize over an LLVM module, since
LLVMModuleAndContext is thread-safe to access.

This should result in a modest compile time boost. The PR also includes
another small optimization: when splitting exported functions and there
is only one function, don't do any splitting (which involves cloning and
roundtripping through bytecode). This should provide a compile time
boost to compilation unit with only one exported function, e.g. `main`.

MODULAR_ORIG_COMMIT_REV_ID: 94827e48a94402e746501092e0daec9288595501