[KGEN] Move `kgen -execute` off of ObjectCompilerLayer (NFC) (#43171)
This is just to see if this even works. We can use StaticArchiveLayer
instead, after emitting the IR straight to object code.

So this apparently works. The only thing to be careful of is
`emitArchive` will modify the module. A lot of code that use
`ObjectCompilerLayer` assume they can read metadata from the module
after passing it to the `ObjectCompilerLayer`.

MODULAR_ORIG_COMMIT_REV_ID: caaaef7fab944aa43fc3a3d534d48e8c432cc035