[mojo] Move `mojo-run` off of ObjectCompilerLayer (#43172)
This removes the only other use of ObjectCompilerLayer. This should
actually speed up everything that uses `mojo run`, since
ObjectCompilerLayer would for some reason produce a clone of the module
before copying it.

This PR also exposes a bug in `splitPerExported`: if two exported
functions depended on another same exported function, then the latter
exported function gets duplicated. The ObjectCompilerLayer was for some
reason tolerant to duplicate external symbols, but the
StaticArchiveLayer is not. This means that this bug could have shown up
in GraphRT, which uses the StaticArchiveLayer.

The fix is to treat exported functions like mutable globals: users of
them have to be kept in the same module. We could split them and mark
the split version of the function as internal, but it's likely that
users of exported functions don't have much extra code in them, so it's
not profitable to split.

MODULAR_ORIG_COMMIT_REV_ID: 28b60973a1da36324cf3f7e20cdf39feaf793211