[KGEN] Delete ObjectCompilerLayer (NFC) (#43173)
`ObjectCompilerLayer::add` is just an unnecessarily fancy way of doing
`emitArchive` and then `StaticArchiveLayer::add`. It's also really
complicated and god knows how many hours Mojo engineers have
collectively lost failing to understand how this thing works.

For example, the ObjectCompilerLayer makes a clone of the module passed
in, which is legacy code from when we make everything a standalone
module. Fine, but when I removed that, for some reason the
ObjectCompilerLayer broke. The symbol resolution algorithm it has (which
was cargo culted from ObjectLayer in LLVM) for some reason requires the
symbol names to be typologically sorted.

Well it's gone now, and it won't be missed!

MODULAR_ORIG_COMMIT_REV_ID: f1929a4b2c58ba00181d00763eadda2f3ed1a21f