[bazel] Fix nested substitutions from default lit cfg (#43210)
Previously nested substitutions didn't work, because we weren't adding
the tool to the PATH. We handled this with one off custom substitutions,
but this meant we dropped any conditional logic from the core litcfg.
Now we automatically add a known set of tools to the PATH and make sure
to overwrite the default substitutions for those (overwriting is
required because it seems like nested substitutions run in order, so the
order of the substitutions list doesn't help).

MAX_INTEGRATION_TESTS_REV_ID: 8fb7b88cb043e0088bf8f3087d8c1e5e233d2bd7