[Serving] [SDK] Refactor tests + add roberta test (#43268)
The PR refactors the standalone test to add onnx and roberta basic
tests. It also adds a roberta case, but it is not enabled by default.
Though it can be uncommented to debug locally.
Currently the tests are not enabled in CI (requires:`manual`) but will
be enabled once we get more confidence in the tests not crashing/failing.
Followup will enable the tests as well as separate the roberta to run
for integration testing.
Reapplying after rebasing on main and fixing `None` mismatch.

---------

Signed-off-by: Rashid Kaleem <rashid@modular.com>
MOJO_MAX_ORIG_REV_ID: eb18d9547922707c0d332f18667d70f59ced3786