[bazel] Do not link TorchRuntimePlugin to Torch libs (#43130)
As described here:
[Internal link]

Do not link libTorchRuntimePlugin to Torch libraries at build time. Our
model is that we will first `dlopen` torch, and then `dlopen` the
plugin. To match that model, the plugin should be built with only the
torch headers and not the actual symbols. The plugin will assume they
are already present when loaded at runtime.

From the build perspective, this requires telling the linker to allow
unresolved symbols when building the plugin. For targets that attempt to
load the plugin, we need to remove it as a dependency and instead
provide it as `data` so that it is available for `dlopen`ing.

In the code, we then need to load torch with `RTLD_GLOBAL` so that the
plugin can see the symbols from torch.

MODULAR_ORIG_COMMIT_REV_ID: f094042957d31c4cb09ac0338b5e2a31d436b907