[Max Driver] Get Engine working with mojo build when executing with device tensors. (#43313)
Context: [Internal link]
We cannot call execute from main thread. We need to call it through an async callback. This is needed because we have checks to verify calling thread is foreign or not and this invariance breaks in mojo build.

However, it was noticed that calling a mojo function pointer from an async callback causes a segmentation fault. So refactor it out to always call from main thread. The crash here is similar to the crash in this issue [Internal link]

MAX_INTEGRATION_TESTS_REV_ID: 99da5cefbd4eabacf9c03c60b1a62cf768c4e3a9