[vscode] Use prettier as formatter (#43432)
Sometimes clang-format does things that hurt my eyes, as it's really not
meant for typescript. I'm switching to prettier which is probably the
best tool for the job.

It can be invoked easily with npm run format and the VSCode prettier
plug-in can do it on the fly.

MODULAR_ORIG_COMMIT_REV_ID: 42033e0b8eae7a054ddb257299efd54f09d4a285