[Kernel] [locks] Split test_lock for better visibility (#43446)
The `test_locks.mojo` has been sporadically failing with limited
visibility. The crash has not been reproduced in local builds. This PR
splits the tests so that each test is in a separate file and we can
confirm that it is not the spin-wait tests, but the locks test that is
failing (or the alternative).

Part of RUNP-220

---------

Signed-off-by: Rashid Kaleem <rashid@modular.com>
RUNTIME_ORIG_REV_ID: b84805c5c0edf22ec65756676d3abdb5a69227ee