[bazel] Add gpu constraints to test targets (#43505)
This adds `exec_compatible_with` in bazel for lit tests that require
running on the GPU. Previously this was only encoded in the test itself
with a requires clause, but for remote execution to know where to
schedule the test, it has to be hoisted into bazel. In the future we can
remove that config from lit.

This uses the gpu tag to infer that the test target requires the gpu,
unless requires_gpu is set to False, or specific test fails are passed
with gpu_only_srcs. This currently leaves tests that don't require the
GPU only running on a CPU machine with remote execution, but in a future
change we will force those to the GPU when a flag is passed.

MAX_INTEGRATION_TESTS_REV_ID: c0bb632e8b30882eb06158cfd75799d2a982841a