[Debugger] Support parsing of pointers to closures (#43520)
The debugger has been crashing when parsing pointer to closures from
debug info. An improper but easy way to fix this is to create signature
decls for these closures. It does the job, but internally mlir complains
when this decl is added to parent structs. I'm fine with this issue for
now because it's benign, as we don't support expr eval yet.

Fixes MOTO-606

MODULAR_ORIG_COMMIT_REV_ID: 949c0e15fa0a2efb24f5cfd6269010b636211415