[mojo-test] Use `mojo run` for unit tests (#43619)
This PR changes the test executor to call out to `mojo run` for unit
tests. This resolves a number of miscompilation errors that occur
because the REPL kernel, which is currently used to run tests, has
compilation issues with Mojo code. Using `mojo run` avoids this, as we
use the same compilation path for tests as we do for actually running
code.

Doctests still run within the REPL kernel. This is unavoidable for
several reasons. Firstly, doctests are written as if they evaluate in
the REPL already (they have top-level constructs that are illegal in
a function, as well as actual runtime code). Secondly, doctests in the
same docstring share context - they are evaluated top-to-bottom.
Replicating this functionality is non-trivial.

There are a few reasons to use a subprocess to call out to `mojo run`
instead of directly interacting with the compiler/runtime:
- By using a subprocess we get process isolation for unit tests. This is
  important to maintain, because if a unit test brings down the process,
  it won't affect other tests.
- `mojo run` already provides interfaces for specifying compilation
  arguments and other such constructs, which we want to allow in tests
  in the near future.
- The complexity of this PR is greatly reduced - we don't have to make
  changes to how tests are dispatched to the executor, nor do we need to
  refactor `mojo run` to extract common functionality.

Fixes MOTO-579.

MODULAR_ORIG_COMMIT_REV_ID: 5c00a4060cfbc10d57e6b322ba8d5fe7ddb45bcd