[mojo-lang][stdlib] improve IntLiteral conversion, support unsigned index
This is in support of conversion for the stdlib UInt type. There were
really two issues: one is that the converter always assumed that index
is signed (like with Int), and the other is that the converter failed
for converting unsigned integers with the high bit set. This fixes them.

Closes [Internal link]
MODULAR_ORIG_COMMIT_REV_ID: fcef1ef4b58ab55c117f8ea9acaef4219aced9db