[mojo-lang] Continued refactoring of CallOperands, NFC (#43817)
This changes variadicKWArgs to use a vector of arguments instead
of the MapVector thing, which is unnecessary.  This also makes
adding a keyword arg to the operand list fail inside of it, instead
of having all the clients do the check.  This is fine because the
parser checks for duplicates already.

MODULAR_ORIG_COMMIT_REV_ID: ae215ede37143a7b06024731f8ff7797d42ea988