[mojo-lang] Fix lifetime inference of ref arguments. (#43822)
After an epic series of refactorings to get the compiler into a
cooperative place,
we can finally fix inference of reference arguments from PValues and
SRValues,
inferring the correct lifetime from them instead of giving up. This will
fix a number
of problems and fixes MOCO-952.

This is complicated because we need to do overload resolution to
determine
which member of an overload set will be used, and different members can
force
different things into memory temporaries because of different reference
arguments.

We handle this by having overload resolution notice when an argument
value
needs to be forced to memory and note this in a new
`argsNeedingLifetimes`
bitset that `OverloadFitness` now carries.  After selecting a candidate,
overload resolution forces that argument to memory and reinfers the
parameters
and lifetimes, then carries on.

This required overload resolution to be able to mutate the argument
list, which
is why I had to do all the work to make CallOperands be owning of the
argument
values etc.

MODULAR_ORIG_COMMIT_REV_ID: 9b0e624249a3f4a631979c61f41a8696dfb52bb4