[KGEN][Custom] Allow JIT'ed `custom` ops canonicalizers to use the mlir bindings  (#43689)
Allows to JIT compile mojo code that uses MLIR C bindings. As we use
this to write `custom` ops
canonicalizers, we need the symbols to be the same as the current
process.

To ensure the MLIR symbols stay alive in the current process, we add a
function autogenerated from the
MLIR C bindings to expose all the symbols we care about.

MODULAR_ORIG_COMMIT_REV_ID: 2109250b5bba650ed969082e52642d72c2579afe