[Support] Move MOGG mlir hasher to support (#43596)
This hasher isn't mogg specific and has some general use outside of MOGG
so this moves it to support. The hasher takes a basic block and returns
a hash for it. It will return the same hash for any basic block which is
equivalent to it.

MODULAR_ORIG_COMMIT_REV_ID: 3519dc398b1e2893995714f810de27c6623033ac