[KGEN] Make elaborator quiescing a turnstile (#43755)
Quiescing tracks the number of outsanding ParamNode tasks, which can
result if we exit due to an error but have overhanging tasks that
weren't completed, and then ensures they have all proceeded to
completion before running any destructors. This was previously written
with a mutex + a task count. We can rewrite this as a turnstile (make
the task count an atomic and start it at 1) to remove the mutex and
simplify some of the code.

MODULAR_ORIG_COMMIT_REV_ID: a7b124bfb4095e2076bf92e907ec30c5f4891f66