[MOGG] Compare function contents when "linking" (#43880)
When we have two functions with the same name in two seperate packages
being loaded by the graph compiler we need to resolve the collision.
Currently we resolve this by really a really basic mechanism of cloning
the function if the signatures don't match. That's a bandaid. Instead
this clones the functions if they don't have the same hash.

This doesn't seem to affect the compile time much

MODULAR_ORIG_COMMIT_REV_ID: b9a047b9278e6e89945b2797c17cd91784ee1114