[KGEN] Switch to explicit dependency tracking in elaborator (#43892)
The elaborator has a "hail mary" cycle detection check because we don't
track all dependencies in the dependency graph, meaning not all nodes in
the expansion graph are reached from the root nodes. This was done
because the cost of dependency tracking was deemed no worthwhile when
recursion was still rare in Mojo code. It's a lot less rare now, and we
need this information to properly handle recursion.

MODULAR_ORIG_COMMIT_REV_ID: c2ea638a08e8aa703165feea96de565efece85dd