[KGEN] Track location in `otherDeps` to give an error message (#43893)
This adds a Location to `otherDeps` which will allow the elaborator to
give an error message when it detects recursion across an interpreter
edge. This should be a significant improvement in QoL when modifying
certain functions that might depend on themselves in an interpreter
context.

MODULAR_ORIG_COMMIT_REV_ID: 7d2a4661944c76fc09dad054ec82c30d51ba2d0b