[KGEN] Synchronize completion of nodes in an SCC in elaborator (#43900)
This PR finally fixes the recursion issue in the elaborator. Now that we
see the full structure of the SCC and can break every edge in the SCC,
we can pile on an addition chain to each ImplNode and
`andThenSyncMoving` with a callback to actually complete all the nodes
in the SCC. This means we release the nodes in the SCC, free to wait on
their callees, which upon completion don't emplace their ParamNode's
chain. Instead, they emplace their SCC chain set by the cycle breaker.

Fixes
[Internal link]
Fixes
[Internal link]
Fixes
[Internal link]

LAYOUT_ORIG_REV_ID: d5a6a78092ecfcd0ec33efe9b65dc2f3ef9d1326