[mojo-stdlib] Optimize the implementation of Variant
Exit early in the parameter loops when we know only 1 if condition can
fire, instead of relying on LLVM to figure this out. Also, ensure
specific functions are called in the parameter domain to ensure we don't
accidentally codegen them. In the future, we should mark them as
consteval or whatever the feature will be called.

MODULAR_ORIG_COMMIT_REV_ID: c00ab62339b59a41bff2c2d85772fe5bcca39930