[KGEN] Introduce a VariantBitcastOp (#44066)
This operation will be used to more safely bitcast a pointer to a
variant to a pointer to one of its element types. This will enable
optimizing in-memory variants more efficiently.

This is part of changing `!kgen.variant` to lower to `{ union, iN }`
post-elaboration.

The op still requires an explicit result type because the variant
element types can be fully parametric.

This op is in the POP dialect instead of the KGEN dialect because it's
meant to be an offramp from KGEN variant to a future union type. We can
probably move the variant ops and type back into POP now that it's not
officially part of the Mojo error handling ABI.

MODULAR_ORIG_COMMIT_REV_ID: f750b0e7cbd910fd5c2ca1c8f1ac5304549be17e