[Mojo][DebugInfo] Hide `__result__` and `__error__` when debugging (#44267)
These variables often appear in the list of variables when debugging and
are very confusing. After adding support for DIFlags in the LLVM
dialect, now I'm adding support for them in mojo.
This is done by marking these variables as `Artificial`. In fact, we
should mark this way every variable or function argument that is
compiler generated and shouldn't be available to the user.

MODULAR_ORIG_COMMIT_REV_ID: dfda0b79483d6e6d328f4662222e0f4803c195a3