[mojo-lang] Reuse Frame For Hot and Cold (#44043)
The only difference between the hot ramp resume and the cold ramp resume
is the arguments that maybe in the frame of the hot resume. But this is
only needed in state 0 which is unreachable from a hold started
coroutine. As a result we bitcast the frame to the larger size only in
state 0 of the resume.

MODULAR_ORIG_COMMIT_REV_ID: 4f6739edc2397ea5466146e2dc30c4a54a65f388