[vscode] Don't load the SDK on startup (#44459)
The VS Code extension's LSP context was loading the SDK (and starting
the LSP server) at startup. This isn't desirable because if you just
install the extension and don't use it, you'll be prompted to install
the SDK on every startup.

The fix is pretty straightforward: load the SDK and start the LSP server
only once you've opened a Mojo file in the workspace somewhere.

A note: we don't have to handle notebooks specifically because they
contain text documents. We'll get an event when a notebook containing
Mojo is opened (or a cell for Mojo is created) through the normal
channels.

MODULAR_ORIG_COMMIT_REV_ID: 2adee23e03afba31a9c9199b8a93e485829ab33d