[graph-api] Add slicing with int, ellipsis, and `:` indices (#44563)
This is just the start of slice and is missing slicing with actual slices, but it gets the base wired through.
It also does not deal with needing to name output dimensions at all.
But at least it gets something working as a base.

This is definitely a ton of wrangling being done in python, but I think it is the quickest path forward.
Eventually, I need to target what a good slice api would be in RMO, but that depends on a number of open questions.

I also need to think of the best way to do negative case hypothesis tests. Currently I feel like each hypothesis tests needs a lot of extra generation tooling.
I think eventually we will have a default library of strategies, but currently it is adhoc.

ref MSDK-664

MAX_GRAPH_API_ORIG_REV_ID: 130125411d07643a22bf85c1572bb8d53e611c27