[mojo-lang] Fix incorrect code in `ExprEmitter::emitStoreToLValue` (#44581)
This code was doing dodgy things where it manually loaded a DLValue.
Upon investigation, this was papering over a problem where it was
trying to avoid redundantly evaluating an expression with side
effects.  The real issue is that we need DLValue::emitStore to
return a BValue for the emitted value that it processes.

Peeling this back revealed problems in tuple lvalue emission which
could cause it to pass an RValue into a consuming getitem impl
multiple times (theoretical though because our tuple doesn't do this).

Fixing this all up simplifies the code and makes it more solid for
other cases, eliminates an unnecessary temporary in the testcase.

While here, simplify the nonmaterializable handling code in the
same function.

MODULAR_ORIG_COMMIT_REV_ID: 4b92d7360697018d1a6b7064bd5ee95aeb3902e9