[mojo-lang] Introduce a new `MBPValue` classification, ~NFC (#44582)
This introduces a new `MBPValue` classification kind to indicate that
the value
is an MValue with parametric mutability. This is considered a BValue
(because you
cannot mutate it, pass it to an inout function etc) but (really
unfortunately) has to
be different from `MBValue` because `MBValue` really does need to be
able to
point to a MLIR reference that is mutable, while indicating that clients
are not
actually allowed to mutate it.

It initially seemed attractive to make IRValue directly match the MLIR
code at all
times: this would require emitting a `lit.ref.immut` op every time (eg)
an
MLValue/MRValue decayed to an MBValue. While this seems attractive, in
practice
we decay almost everything to an MBValue and often ignore it, so this
would
generate tremendous numbers of casts.

This leaves parametric mutability values in a funny place: they cannot
be
represented as an MLValue because the user isn't allowed to assign into
them,
they cannot be an MBValue because call emission would then force them to
an
immutable ref when passing (losing the parametric mutability) and of
course
cannot be an MRValue because they aren't owning.  As such, we need a new
value kind.

The good news is that this isn't actually all that invasive and directly
modeling
this is required for downstream patches.

MODULAR_ORIG_COMMIT_REV_ID: 690961e54fe198809a428d06488218a14fe29235