[KGEN] Make DiagnosticHandler filter in Elaborator. (#43591)
The `DiagnosticHandler` used in IREvaluator is created for each
invocation of `evaluateCompileAssembly`. `evaluateCompileAssembly` can
be invoke in parallel in multiple threads and have subinvocations. While
the `DiagnosticEngine` is thread safe to which the handler is added to,
the handler needs to filter the `Diagnostic` that it is handling for the
corresponding `evaluateCompileAssembly` and pass the others back to the
DiagnosticsEngine to be handled by the right handler.

The solution added here is to compare the handler's invocation thread's
id against the thread id when it is created (which should be associated
with the ``evaluateCompileAssembly` invocation) for filtering out
`diag`s that are created by other threads hence no need to handled. And
attach some metadata to indicate if the handler is being called when
`emit`ing the error (when the `diag` should be processed or just being
called when the `drag` is being sent back to the engine and no need to
handle it).

This PR relies on an upstream change to work:
https://github.com/llvm/llvm-project/pull/99398

MODULAR_ORIG_COMMIT_REV_ID: 9985c44926974bede5fa3d0cd765fe4270d5ca9d