[mojo-lsp] Simplify a test in `CodeBlockTest.cpp` (#44721)
Clarify the code blocks in this test to rename the variable from `value`
to some other identifier that isn't "value". This helps distinguish the
completion result entry expected has nothing to do with the identifier
used in the program. In this case, it's expecting the label "value"
because the type of the variable in the code block, `Int`, has a data
member whose identifier is "value".

MODULAR_ORIG_COMMIT_REV_ID: 959a1eaa89bcc1fd10b31de0628c92b99e01725b