[External] [stdlib] Add a `unsafe_assume_initialized` constructor to `InlineArray` (#42256)
This is a follow-up on
https://github.com/modularml/mojo/pull/2965#discussion_r1648153449

The idea here is to guide the user in the unsafe lands of programming.
Such that even when manipulating uninitialized memory, the assumptions
about the data are clear.

The next step is to give a new constructor so that doing
```
InlineArray[UnsafeMaybeUninitialized[String], 3](unsafe_uninitialized=True)
```

become a bit more user-friendly. We'd like to make it impossible to use
`unsafe_uninitialized` without working with `UnsafeMaybeUninitialized`.
This is for a next PR.

@ConnorGray for review as you were the one to follow the work in this
area.

Co-authored-by: Gabriel de Marmiesse <gabriel.demarmiesse@datadoghq.com>
Closes modularml/mojo#3119
MODULAR_ORIG_COMMIT_REV_ID: 9a57d97dbd4608f3cf20510137cb4cc2dcb64178