[mojo-lang] Make computed lvalue writebacks transfer the temp. (#44728)
This changes computed LValue writebacks to /transfer/ the value that
is get'ed back into the setter for the property/index.  This composes
well with setters that take the value owned, and makes sure we diagnose
attempts to accidentally extend the lifetime of the temporary.

MODULAR_ORIG_COMMIT_REV_ID: 3d888a485c9c64903ed7ec7653798debb1f6d186