[GraphAPI] Remove Graph._output_types (#44747)
This was causing MyPy issues, and after discussion with Brendan H, it
seems like we do not actually use it anywhere. All the actual use sites
for Graph leave output types empty, and it is filled in when `output()`
is called. So we don't actually need to store this.

MAX_GRAPH_API_ORIG_REV_ID: e0496085143054d86bc65978ec3987f51a201ee2