[KGEN] Tighten verification of call ops and func ops (#44774)
This PR tightens the verifiers for call operations to check operand and
call argument types. It also checks function body argument types against
the function signature.

This revealed a bug in ResolveCompilerPromises adding operands in the
wrong order when recursion happens... And fixes it.

This also revealed bugs in the graph compiler generating invalid IR in
its fusion passes. The passes are tightly wound and it's not clear to me
where the fix is supposed to go (invalid replacements across parameter
domains), so I massively hacked in a walk that just inserts rebinds when
necessary.

MODULAR_ORIG_COMMIT_REV_ID: 7c1073897756a6c1f50dfc0182c3425758f1fb5c