[Graph API] Make Dim an alias for int | str  (#44813)
Fixes [MSDK-749: \[Graph API\] Make python `Dim` be `int | str` instead
of a
[Internal link]

We've run into a _bunch_ of cases where we try to
treat ints or strings as a `Dim` directly, and it's routinely
a footgun. This PR

- Makes `Dim` an `int | str | mlir.Attribute`
- Pushes this change through Graph API
- Improves some testing of `reshape` which started failing
  during these changes.

It also, maybe controversially, makes it possible for the Graph API
to return algebraic dimensions back to the user. This was required
to not change some reshape behavior along with this change.

MAX_GRAPH_API_ORIG_REV_ID: e5151058122b3a6692369f036b342a026302eb5a