[Kernels][KVCache] Create a fused QKV proj and RoPE kernel (#44865)
We want to have both a fused QKV implementation, and a RoPE kernel for
the llama family of models. There is a problem when these are combined,
in that the QKV kernel can't express both a NDBuffer and `mo.opaque`
output. This issue should address that, in the meanwhile we manually
fuse these kernels together GRA-863

NN_ORIG_REV_ID: 6734c354b1438ab6e35cde6a4bc3b0e4ff76374e