[KGEN] Handle private asan symbol during llvm splitting (#44091)
`mojo --sanitize=address` introduces a few private global constants
(`___asan_gen_.0`, `___asan_gen_.1`, ...) during llvm optimization. When
we split an llvm optimized module with these symbols, they are not
handled properly (we are only handling `internal` linkages and marking
them `weak` to make the splitting work).

The bigger issue here is that during splitting, we mark global constant
value as `canBeSplit` like a function. So during splitting, they are not
copied properly to each of their own user's module because `canBeSplit`
means that it is not a dependency to the user (like a callee can be in
its own module instead going to the caller's module). And we were
fudging the linkage type from `internal` to `weak`, so that even there
are multiple modules use the constant global variable, they can still
find where it is as a weak linkage. This is obviously dangerous.

Since these variables are constant (and we don't have coro anymore), we
should be able to just copy them to each user's own module without
changing the linkage type.

---------

Co-authored-by: Keith Smiley <keithbsmiley@gmail.com>
MODULAR_ORIG_COMMIT_REV_ID: de89d84a48a141b50b9637a0087eea8ef2b27de5