[mojo][test] Fix nightlies and add a self test (#44993)
Nightlies are broken because some necessary logic was being inside an
assert, and release SDKs don't execute asserts, which resulted in no
tests being able to be executed.
This wasn't caught because we didn't have a self-test to be executed as
part of the SDK generation, so I'm adding one for that.
I also changed all the asserts to use either report_fatal_error or just
to return an Error object.

MODULAR_ORIG_COMMIT_REV_ID: a83f2079238ce2d330f0bfc0bc48575dc0815048