[KGEN][Custom] Move custom op params preservation to its own pass (#44596)
Previously, the `register-custom-ops` pass was wrapping custom ops
parameters in a
`kgen.preserved` attribute, so the types would not get lowered further.
However, this is causing
problem in the elaborator, so this feature is now moved to its own pass,
that is run right after the elaborator.

This PR also modify an integration test that was previously failing
because of this `kgen.preserved` attribute
being added too early on.

MODULAR_ORIG_COMMIT_REV_ID: 27d592857a8fe84f98901e4205422dbf8adcb427