[mojo-test] Invoke entrypoint in the test executor (#44366)
There's not much left of the test executor, excepting REPL work, but
it's a non-trivial refactor to remove it and have `mojo test` drive test
execution, so we'll leave it for later.

One gotcha here: escape sequences in test IDs. If your test function
contains a character like `.`, which is not legal without being enclosed
in backticks, it will be escaped during test collection. This causes us
some problems for the test entrypoint, because we'd have to
double-escape that sequence to include it in a string literal. We take a
different approach and undo the escape sequence, producing a test ID
that won't cause problems when pasted into a string literal.

I also added an assertion to the test entrypoint that will fail if _no_
test executes (e.g. you passed the wrong test ID). This should never be
hit in production but it is a useful sanity check for us.

MODULAR_ORIG_COMMIT_REV_ID: aed00731328c4a2057bb48c1dc71b38854fad52e