[Kernels][MOGG] Tack on arch to `target` parameter in kernels (#44901)
With this change, the `target` string in the kernels contains the device
arch (i.e. cuda-sm_8x). Ideally we would like to pass in the whole
``__mlir_type.`!kgen.target`` as a parameter for the kernels to use, but
the string can temporarily be used to dispatch on the target at compile
time (along with [Internal link]

Fixes GRA-808.

---------

Co-authored-by: Ian Tramble <ian@modular.com>
NN_ORIG_REV_ID: f952c3737f50d6bce724bfa3d689b8451b2dbea8