[External] [stdlib] Use memcpy in `List._realloc` for x13.2 speedups in some benchmarks (#44680)
[External] [stdlib] Use memcpy in `List._realloc` for x13.2 speedups in
some benchmarks

Currently, when re-allocating a list, elements are moved one by one.
While it is correct, when one knows that `T` is a trivial type, we an
use `memcpy` and get massive speedups in some benchmarks.

Since the compiler currently doesn't expose this information about a
type, we rely on the information being given by the caller. If the
caller doesn't give this information, we fallback to a normal loop which
will move each element.

We fully expect this parameter to be temporary and to go away in the
future, when the compiler exposes more information about types.

This is extremely useful when working with the `String` type, as moving
bytes one by one is quite slow (I noticed it when looking at profiling
data and the assembly code when working on SSO).

Here are the speedups in some benchmarks (I only use the hint in the
`String` struct, as the PR shows). Here is the hardware:

Intel(R) Core(TM) i7-10700KF CPU @ 3.80GHz, 16gb ram with WSL 2, windows
11 pro 21H2

* `str(x: Int)` on the first 1000000 integers: 185820us -> 186953us
(x1.0)
* Adding one char to a string in a loop 30000 times 159887us -> 12103us
(x13.2)
* Parsing a json list of 5000 integers: 345455us -> 342679us (x1.0)
* Call `str(x: List[Int])` on a list of 10000 integers: 205393us ->
17321us (x11.6)

Currently this optimization is applied to `String` automatically, but we
can expect speedups in the future when the compiler gives the info, and
we'll then get speedups for types like `List[Int]`, `List[SIMD[...]]`
etc...

Co-authored-by: Gabriel de Marmiesse <gabriel.demarmiesse@datadoghq.com>
Closes modularml/mojo#3349
MODULAR_ORIG_COMMIT_REV_ID: 83c9c8cd85acf6df95c1f7e82cf0a00b3a0d256a