[KGEN] Fix `SignatureType::getSpecializedSignature` (#45161)
This fixes a bug in `SignatureType::getSpecializedSignature`
handling partial specialization of signatures with dependent
parameter types.  This is part of #44966 but split out because
it is independent.  Unfortunately I don't have a testcase for
this because it is so cornercase, but it is tested by that
forthcoming patch.

MODULAR_ORIG_COMMIT_REV_ID: aaacffc7d4037894a9ab0c93d1a852d513f6f1c2