[KGEN] Struct type-value as symbol (#44831)
A mojo source struct can be represented with a `kgen.struct.generator`
op, which is like a `kgen.generator` op, except it represents a
templated value (type-value), instead of a templated function.

It's arguable whether we still need a `inst_struct` operator since we
have a dedicated `typeref` attr that refers to type-value symbols. The
benefit of an explicit instantiation operator is it creates an
opportunity for the parameter evaluator to hand control back to the
elaborator for "evaluation" of the expression, at which point it's able
to perform this instantiation. Otherwise we'd need a separate hook for
instantiating references (which could work too).

MODULAR_ORIG_COMMIT_REV_ID: 4fb9a0f2a0ceb5840fb98fa5f56fd386644721db