[bazel] Delete standalone targets and run_tool (#45163)
The run_tool rule was for mapping paths of runtime dependencies to where
they exist in the bazel build directory. With the addition of a
`//:install` target, these paths can now be discovered by reading the
`modular.cfg` file that's written there.

This makes it a bit more acceptable to not fully enumerate all potential
dependencies that could exist for a target. For example mojo's
dependencies varied wildly depending on what you were building.

Instead we now build a reasonable set of dependencies by adding them to
each CLI's `data` attribute, and we assume that any others used will be
rebuilt by the user and set in the modular.cfg.

This change still injects a small shim that sets the right env, but this
eliminates the need for remapping a lot of relative paths. With this
change we also now point to the autovenv for python deps instead of in
bazel.

This change generally makes running binaries less hermetic than before,
which might lead to issues, but given that folks can also run the
binaries directly outside of bazel, it's similar to that.

MODULAR_ORIG_COMMIT_REV_ID: d0f021432621772c6b3d210f1f03a7dd351a3ac8