[PT][SDK][Engine] Do not implicitly convert data types in `Model.execute` (#45283)
It is unclear why we allowed implicit data types conversions for tensor
inputs, since this will usually trigger an allocation. This was also not
applied for nested input types (e.g. list of tensors), so it was
inconsistent. This patch simply removes this behavior, and fixes tests
that relied on it.

MAX_INTEGRATION_TESTS_REV_ID: 9b5ee3192964b258a028d154c9093ca3b5bd0867