[KGEN] Elaborate struct generator (#44840)
Elaborate StructGeneratorOps in the elaborator.

- Many places that previously only expected a GeneratorOp now expects
any GeneratorOpInterface op.
- Elaboration of StructInstantiate is special in that it does not
require the "callee" StructGeneratorOp to finish elaborating. It only
requires a reference to the final concrete symbol, which can be known
when elaboration initially gets dispatched. This allows us to have
recursive struct definitions without incurring the cost of explicit
cycle-breaking. We only require that the types eventually get elaborated
so that any references can be resolved.
- To ensure types eventually finish elaboration (e.g. in case they get
blocked by a non-type cycle), we introduce a new kind of dependency in
each ImplNode: an "eventualDeps" list holds all the nodes that
eventually need to finish (we don't care about their results, we just
care that they exist). This ensures we can still reach all nodes from
the primary generators. To make completion checking faster, we
additionally track an atomic that counts the number of unfinished type
nodes. This has to go to zero before we're considered done.

Next PR creates `StructInstanceOp` so the result of elaborating a
StructGeneratorOp is no longer a FuncOp.

MODULAR_ORIG_COMMIT_REV_ID: 172dfb295871165561bc464796b659daac2b979a