[Kernels][GPU] Make sure that error is presented if driver is too old (#45378)
If the CUDA driver version is too old then we cannot compile for that
and before presented the user with an INVALID_PTX error. This catches
the error a bit earlier and makes sure that a human error is presented
to the user.

Closes KERN-813

GPU_ORIG_REV_ID: 6e8e0bcf82e36960d7e0743f9057126b4a964bfc