[mojo-lang] Fix lifetime mutability tracking. ~NFC (#45464)
This fixes logic in the lifetime finder to handle imm cast's
correctly.  Consider a function call like:

  `foo(Slice[LifetimeImmCast(some_mut_lifetime)])`

It is important for us to see that foo is an access to
`some_mut_lifetime`, but low-level lifetime finding logic shouldn't
discard the cast around it: some clients care and others don't, so
let the client discard the imm cast if necessary.

This doesn't cause any behavior change with the logic in tree,
but will when exclusivity checking comes in: we don't want to treat
this as a mutable use of the lifetime.

MODULAR_ORIG_COMMIT_REV_ID: a7e58bc3b0b8d7d515112cac35ea16c343db7d0a