[mojo-lang] Reimplement argconvention result ref check. (#45486)
This implements a new solution to the problem reported in
[Internal link]  Instead of
rebinding generic argument references to have a
`InvalidRefLifetimeAttr` lifetime, check for the problem during
result emission.

Here's the intuition: we can allow binding references of all kinds
to borrowed/inout arguments - even if they may expand to a
register_passable type - so long as the reference isn't returned
from the function.  The problem with returning the reference is
actually a lifetime problem due to how ArgConvention lowering
works.

The result of this is a lot less code, the elimination of
InvalidRefLifetimeAttr, less IR for the compiler to push around,
and fixes the exclusivity checker's handling of these arguments
(as now diagnosed in call-emission-errors.mojo).

MODULAR_ORIG_COMMIT_REV_ID: 4876584aad6e02e07b3a326aab36b4dfb67f2617