[MO][MAX Driver] Add target info options and populate from MAX Driver (#45441)
Previously we relied on the graph compiler to detect the GPU compute
capability by calling into CUDA. In order to remove this CUDA dependency
from the graph compiler, this PR adds 2 compile options to specify the triple
and architecture for the defaultDevice and sets those options from the
MAX Driver. These options are added to a target_info attribute on the MO
graph, which non-host ops use to set the `target` parameter going into
the kernel with the correct architecture.

These options can be removed when the device attributes are added to the
MO graph explicitly.

Fixes GRA-903.

MODULAR_ORIG_COMMIT_REV_ID: 8026b560776865f6010028d724b43993aef4bd68