[KGEN][CMake] Deprecate `validate_mojo_dependencies` (#45582)
In cmake you can accidently not include dependencies and things will generally build.

This is because cmake carries state between invocations of build (e.g. you might not say you want foolib for barlib but built foolib accidentally already -- there is now a race condition in whether building foolib succeeds if you call `build` on a fresh build).

To get around this we have a script called
`validate_mojo_dependencies.py` which tries to get around this by examining all the import statements in a mojo library. Unfortunately, it doesn't know about comments. For example let's say we have a file in a public facing API called `bar`:

```

"""
Example of use:

from bar import bar_obj

cheem(bar_obj)
...
"""

import math
...
```

The script will see the `import bar` line and complain the cmake file needs itself as a dependency!

Of course the error is the commented section does not impose an actual dependency.

We could fix it, but all of this stuff is not useful anymore with the advent of bazel and the fact cmake is in the process of deprecation.

Towards MSDK-827

MODULAR_ORIG_COMMIT_REV_ID: baf8c47aeba56f1a032d4f81ac1b56d2d22100e7